Transcoder Service/requirements.txt
fastapi
boto3
pydantic_settings
python-dotenv
requests